Patch, post, delete enrollments

+++ OnlyCreateDatabase/Controllers/EnrollmentController.cs
@@ -13,28 +13,39 @@ namespace OnlyCreateDatabase.Controllers
     {
         private readonly ILogger<EnrollmentController> logger;
         private readonly IEnrollmentService enrollmentService;
+        private readonly ICourseService courseService;
 
-        public EnrollmentController(ILogger<EnrollmentController> _logger, IEnrollmentService _enrollmentService)
+        public EnrollmentController(ILogger<EnrollmentController> _logger, IEnrollmentService _enrollmentService, ICourseService _courseService)
         {
             logger = _logger;
             enrollmentService = _enrollmentService;
+            courseService = _courseService;
         }
 
-        [HttpGet("InCourse/{courseId}")]
+        [HttpGet("InCourse/{courseId}")] //Userzy w kursie
         public IActionResult InCourse([FromRoute] int courseId)
         {
             var model = enrollmentService.UsersInCourse(courseId);
             return Ok(model);
         }
 
-        [HttpGet("NotInCourseYet/{courseId}")]
+        [HttpGet("NotInCourseYet/{courseId}")] //User, którzy są zapisani, ale jeszcze nie zakacepotwani
         public IActionResult NotInCourse([FromRoute] int courseId)
         {
-            var model = enrollmentService.UsersNotInCourseYet(courseId);
-            return Ok(model);
+            var teacherId = HttpContext.User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+
+            if (teacherId != null && courseService.IsUserOwnerCourse(int.Parse(teacherId), courseId))
+            {
+                var model = enrollmentService.UsersNotInCourseYet(courseId);
+                return Ok(model);
+            }
+            else
+            {
+                return BadRequest("You are not owner of this course!");
+            }
         }
 
-        [HttpGet("InvitedCourse")]
+        [HttpGet("InvitedCourse")] //Kursy usera do których został zaproszony
         public IActionResult InvitedCourse()
         {
             var userId = HttpContext.User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
@@ -49,7 +60,7 @@ namespace OnlyCreateDatabase.Controllers
             } 
         }
 
-        [HttpGet("MyCourses")]
+        [HttpGet("MyCourses")] //Kursy zalogowanego usera
         public IActionResult MyCourses()
         {
             var userId = HttpContext.User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
@@ -65,13 +76,10 @@ namespace OnlyCreateDatabase.Controllers
             
         }
 
-        
-
         [HttpPost("JoinCourse/{courseId}")]
         public IActionResult JoinToCourse([FromRoute] int courseId)
         {
             var userId = HttpContext.User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
-            var userName = HttpContext.User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Name)?.Value;
             if(enrollmentService.IsAlreadyInCourse(int.Parse(userId), courseId) == false)
             {
                 enrollmentService.JoinCourse(int.Parse(userId), courseId);
@@ -81,21 +89,86 @@ namespace OnlyCreateDatabase.Controllers
             {
                 return BadRequest("User jest już zapisany!");
             }
-            
         }
 
-        [HttpPatch("AcceptJoin/{courseId}")]
-        public IActionResult AcceptToCourse(List<int> usersId, [FromRoute] int courseId)
+        [HttpPatch("AcceptJoin/{courseId}")] //Teacher
+        public IActionResult AcceptToCourse([FromQuery] int[] usersId, [FromRoute] int courseId)
         {
-            enrollmentService.AcceptJoin(usersId, courseId);
-            return Ok();
+            var teacherId = HttpContext.User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+            if (teacherId != null && courseService.IsUserOwnerCourse(int.Parse(teacherId), courseId))
+            {
+                enrollmentService.AcceptJoin(usersId, courseId);
+                return Ok();
+            }
+            else
+            {
+                return BadRequest("You are not owner of this course!");
+            }
         }
 
-        [HttpDelete("RemoveUser/{courseId}")]
-        public IActionResult RemoveUserFromCourse(List<int> usersId, [FromRoute] int courseId)
+        [HttpPatch("DeleteJoin/{courseId}")] //Teacher
+        public IActionResult DeleteJoin([FromQuery] int[] usersId, [FromRoute] int courseId)
         {
-            enrollmentService.RemoveUserFromCourse(usersId, courseId);
-            return Ok();
+            var teacherId = HttpContext.User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+            if (teacherId != null && courseService.IsUserOwnerCourse(int.Parse(teacherId), courseId))
+            {
+                enrollmentService.DeleteJoin(usersId, courseId);
+                return Ok();
+            }
+            else
+            {
+                return BadRequest("You are not owner of this course!");
+            }
         }
+
+        [HttpPatch("AcceptInvite/{courseId}")]
+        public IActionResult AcceptInvite([FromRoute] int courseId)
+        {
+            var userId = HttpContext.User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+            if (userId != null)
+            {
+                enrollmentService.AcceptInvite(int.Parse(userId), courseId);
+                return Ok();
+            }
+            else
+            {
+                return BadRequest("Błąd w przekazaniu tokena!");
+            }
+        }
+
+        [HttpDelete("DeleteInvite/{courseId}")]
+        public IActionResult DeleteInvite([FromRoute] int courseId)
+        {
+            var userId = HttpContext.User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+            if (userId != null)
+            {
+                logger.LogWarning(userId.ToString());
+                enrollmentService.DeleteInvite(int.Parse(userId), courseId);
+                return Ok();
+            }
+            else
+            {
+                return BadRequest("User jest już zapisany!");
+            }
+        }
+
+        [HttpDelete("RemoveUserFromCourse/{courseId}")]
+        public IActionResult RemoveUserFromCourse([FromQuery] int[] usersId, [FromRoute] int courseId)
+        {
+            var teacherId = HttpContext.User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+            if (teacherId != null && courseService.IsUserOwnerCourse(int.Parse(teacherId), courseId))
+            {
+                enrollmentService.RemoveUserFromCourse(usersId, courseId);
+                return Ok("Usunięto!");
+            }
+            else
+            {
+                return BadRequest("You are not owner of this course!");
+            }
+        }
+
+        
+
+        
     }
 }
